Enable native AOT deployment for daemon and client
This patch enables native AOT deployment for both the daemon and
client projects.  This produces an app that's self-contained and that
has been ahead-of-time (AOT) compiled to native code.

This means that the apps do not require an installed .NET runtime, has
a very fast startup time and smaller memory footprints.

* https://learn.microsoft.com/en-us/dotnet/core/deploying/native-aot/
* https://learn.microsoft.com/en-us/dotnet/core/deploying/trimming/trim-self-contained

This patch also replace the protobuf-net NuGet package with
Google.Protobuf.  The reason is that the protobuf-net package does not
support native AOT deployment, but the Google package does.

This means that we need to generate the C# code using the protoc
compiler.  Just like we do in the X11/Wayland Linux C implementation.

* https://github.com/altenstedt/flextime-x11

This patch also change the way we use the HttpClient *AsJson methods
to support trimming.

* https://learn.microsoft.com/en-us/dotnet/standard/serialization/system-text-json/source-generation-modes?pivots=dotnet-8-0#metadata-based-mode
* https://learn.microsoft.com/en-us/dotnet/standard/serialization/system-text-json/source-generation?pivots=dotnet-8-0